+++ src/app/inobeta-ui/ui/material-forms/controls/autocomplete.spec.ts
@@ -1,0 +1,78 @@
+import { async, TestBed } from '@angular/core/testing';
+
+import { IbToolTestModule } from '../../../tools';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { IbDynamicFormsModule } from '../../forms';
+import { MatAutocompleteModule, MatFormFieldModule, MatIconModule, MatOptionModule, MatSelectModule } from '@angular/material';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { IbMatAutocompleteComponent } from './autocomplete';
+
+
+
+describe('IbMatAutocompleteComponent', () => {
+  let component: IbMatAutocompleteComponent;
+  let formBuilder: FormBuilder;
+  let control: FormControl;
+  let form: FormGroup;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [
+        IbMatAutocompleteComponent,
+      ],
+      imports: [
+        IbToolTestModule,
+        CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        IbDynamicFormsModule,
+        MatFormFieldModule,
+        MatOptionModule,
+        MatAutocompleteModule,
+        NoopAnimationsModule,
+        FlexLayoutModule,
+        MatIconModule
+      ]
+    })
+    .compileComponents();
+
+    formBuilder = TestBed.get(FormBuilder);
+    control = new FormControl('control');
+    form = formBuilder.group(control);
+  }));
+
+  beforeEach(() => {
+    const componentRef = TestBed.createComponent(IbMatAutocompleteComponent);
+    component = componentRef.componentInstance;
+    component.data = {
+      self: control,
+      base: {
+        key: 'control',
+        label: 'control',
+        width: '33.3%',
+        options: [
+          { value: 'value1' },
+          { value: 'value2' },
+          { value: 'value3' },
+          { value: 'value4' }
+        ]
+      },
+      form,
+      hasError: null,
+      formControlErrors: null
+    };
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+
+
+  it('should search', () => {
+    component.onSearchChange('1', component.data.base.options);
+    expect(component.autocompleteFiltered.length).toBe(1);
+  });
+
+});