+++ src/app/inobeta-ui/ui/material-forms/index.ts
@@ -14,3 +14,4 @@ export * from './controls/textbox';
 export * from './controls/textarea';
 export * from './controls/button';
 export * from './controls/padding';
+export * from './controls/slide-toggle';